+++ src/components/hero/types.ts
@@ -1,4 +1,11 @@
-import { Crumb } from './breadcrumbs';
+export interface Crumb {
+    text: string;
+    url?: string;
+}
+
+export interface BreadcrumbsProps {
+    crumbs: Crumb[];
+}
 
 export interface CTA {
     text: string;
