--- src/interfaces/components/breadcrumbs.ts
@@ -1,8 +1,0 @@
-export interface Crumb {
-    text: string;
-    url?: string;
-}
-
-export interface BreadcrumbsProps {
-    crumbs: Crumb[];
-}
